# api/openapi.yaml
openapi: 3.0.3
info:
  title: Music info
  version: 0.0.1
paths:
...
  /verse/{id}:
    get:
      parameters:
        - name: id
          in: path
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: true
          schema:
            type: integer
        - name: size
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetVerseResponse'
        '400':
          description: Bad request
        '500':
          description: Internal server error
components:
  schemas:
...
    GetVerseResponse:
      required:
        - text
      type: object
      properties:
        text:
          type: array
          items:
            type: string
            example: ля
...